hermes-backend/requirements.txt
fastapi
uvicorn
polars
requests
python-dotenv
pyarrow
aiohttp
numpy
pydantic-settings
ruff
mypy
types-requests
pytest
pytest-cov
